Ajustes variados
- Se cambio ligeramente la comprobación del estado del botón
- Se cambio el calculo del desplazamiento en el efecto de imagen infinita, corrigiendo así un error que aumentaba significativamente el uso de recursos.
- Se actualizo el código de Arduino

+++ efectos.pde
@@ -25,8 +25,8 @@ class Efecto {
   }
   
   void imagen_infinita(PImage imagen, float pos_y, float velocidad) {
-    float desplazamiento = (frameCount * velocidad) % imagen.width;
-    for (float pos_x = -desplazamiento; pos_x < width; pos_x += imagen.width) {
+    int desplazamiento = int(frameCount * velocidad) % imagen.width;
+    for (int pos_x = -desplazamiento; pos_x < width; pos_x += imagen.width) {
       image(imagen, pos_x, pos_y);
     }
   }
